+++ service-base/src/main/java/org/eclipse/hono/service/http/HttpServiceBase.java
@@ -21,7 +21,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 
-import org.eclipse.hono.config.ServiceConfigProperties;
 import org.eclipse.hono.service.AbstractServiceBase;
 import org.eclipse.hono.util.Constants;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -43,7 +42,7 @@ import io.vertx.ext.web.handler.AuthenticationHandler;
  *
  * @param <T> The type of configuration properties used by this service.
  */
-public abstract class HttpServiceBase<T extends ServiceConfigProperties> extends AbstractServiceBase<T> {
+public abstract class HttpServiceBase<T extends HttpServiceConfigProperties> extends AbstractServiceBase<T> {
 
     private final Map<String, HttpEndpoint> endpoints = new HashMap<>();
 
@@ -271,8 +270,10 @@ public abstract class HttpServiceBase<T extends ServiceConfigProperties> extends
     protected HttpServerOptions getHttpServerOptions() {
 
         final HttpServerOptions options = new HttpServerOptions();
-        options.setHost(getConfig().getBindAddress()).setPort(getConfig().getPort(getPortDefaultValue()))
-                .setMaxChunkSize(4096);
+        options.setHost(getConfig().getBindAddress())
+                .setPort(getConfig().getPort(getPortDefaultValue()))
+                .setMaxChunkSize(4096)
+                .setIdleTimeout(getConfig().getIdleTimeout());
         addTlsKeyCertOptions(options);
         addTlsTrustOptions(options);
         return options;
@@ -292,7 +293,8 @@ public abstract class HttpServiceBase<T extends ServiceConfigProperties> extends
 
         final HttpServerOptions options = new HttpServerOptions();
         options.setHost(getConfig().getInsecurePortBindAddress())
-                .setPort(getConfig().getInsecurePort(getInsecurePortDefaultValue())).setMaxChunkSize(4096);
+                .setPort(getConfig().getInsecurePort(getInsecurePortDefaultValue())).setMaxChunkSize(4096)
+                .setIdleTimeout(getConfig().getIdleTimeout());
         return options;
     }